ARFF File Save Format & Cursor Addition
The program now saves all run coordinates to an ARFF file, and there was a mouse cursor added to the screen to show the user where they are drawing. A later addition will have the cursor moving at all times on the screen, no matter if the user is drawing.

More Additions: Two methods - one to offset coordinates to start at (0, 0) and method to transform signatures that are too big to smaller sizes. Will continue refining after machine learning research.