Digital Signature Recognition Process
The program now saves all run coordinates to a separate file in the root java folder you have assigned, and conducts the vector method on the signature itself. Then, that information is fed into WEKA to compare to a test ARFF file.

Note: Number of cells can be adjusted as desired by changing the size of each cell check and the number of attributes in the ARFF file creation format.